sql: dont explicitly test datastore, instead use postgres to verify
