+++ sql/postgres/postgres_test.go
@@ -1,6 +1,10 @@
 package postgres
 
 import (
+	"bytes"
+	"crypto/rand"
+	"fmt"
+	"sort"
 	"strings"
 	"testing"
 
@@ -151,7 +155,7 @@ func TestSetDefaultOptions(t *testing.T) {
 
 func TestQuery(t *testing.T) {
 	d, done := newDS(t)
-	defer done(t)
+	defer done()
 
 	addTestCases(t, d, testcases)
 
@@ -243,10 +247,10 @@ func TestQuery(t *testing.T) {
 
 func TestHas(t *testing.T) {
 	d, done := newDS(t)
-	defer done(t)
+	defer done()
 	addTestCases(t, d, testcases)
 
-	has, err := d.Has(datastore.NewKey("/a/b/c"))
+	has, err := d.Has(ds.NewKey("/a/b/c"))
 	if err != nil {
 		t.Error(err)
 	}
@@ -255,7 +259,7 @@ func TestHas(t *testing.T) {
 		t.Error("Key should be found")
 	}
 
-	has, err = d.Has(datastore.NewKey("/a/b/c/d"))
+	has, err = d.Has(ds.NewKey("/a/b/c/d"))
 	if err != nil {
 		t.Error(err)
 	}
@@ -267,10 +271,10 @@ func TestHas(t *testing.T) {
 
 func TestNotExistGet(t *testing.T) {
 	d, done := newDS(t)
-	defer done(t)
+	defer done()
 	addTestCases(t, d, testcases)
 
-	has, err := d.Has(datastore.NewKey("/a/b/c/d"))
+	has, err := d.Has(ds.NewKey("/a/b/c/d"))
 	if err != nil {
 		t.Error(err)
 	}
@@ -279,12 +283,12 @@ func TestNotExistGet(t *testing.T) {
 		t.Error("Key should not be found")
 	}
 
-	val, err := d.Get(datastore.NewKey("/a/b/c/d"))
+	val, err := d.Get(ds.NewKey("/a/b/c/d"))
 	if val != nil {
 		t.Error("Key should not be found")
 	}
 
-	if err != datastore.ErrNotFound {
+	if err != ds.ErrNotFound {
 		t.Error("Error was not set to ds.ErrNotFound")
 		if err != nil {
 			t.Error(err)
@@ -294,10 +298,10 @@ func TestNotExistGet(t *testing.T) {
 
 func TestDelete(t *testing.T) {
 	d, done := newDS(t)
-	defer done(t)
+	defer done()
 	addTestCases(t, d, testcases)
 
-	has, err := d.Has(datastore.NewKey("/a/b/c"))
+	has, err := d.Has(ds.NewKey("/a/b/c"))
 	if err != nil {
 		t.Error(err)
 	}
@@ -305,12 +309,12 @@ func TestDelete(t *testing.T) {
 		t.Error("Key should be found")
 	}
 
-	err = d.Delete(datastore.NewKey("/a/b/c"))
+	err = d.Delete(ds.NewKey("/a/b/c"))
 	if err != nil {
 		t.Error(err)
 	}
 
-	has, err = d.Has(datastore.NewKey("/a/b/c"))
+	has, err = d.Has(ds.NewKey("/a/b/c"))
 	if err != nil {
 		t.Error(err)
 	}
@@ -321,14 +325,14 @@ func TestDelete(t *testing.T) {
 
 func TestGetEmpty(t *testing.T) {
 	d, done := newDS(t)
-	defer done(t)
+	defer done()
 
-	err := d.Put(datastore.NewKey("/a"), []byte{})
+	err := d.Put(ds.NewKey("/a"), []byte{})
 	if err != nil {
 		t.Error(err)
 	}
 
-	v, err := d.Get(datastore.NewKey("/a"))
+	v, err := d.Get(ds.NewKey("/a"))
 	if err != nil {
 		t.Error(err)
 	}
@@ -340,7 +344,7 @@ func TestGetEmpty(t *testing.T) {
 
 func TestBatching(t *testing.T) {
 	d, done := newDS(t)
-	defer done(t)
+	defer done()
 
 	b, err := d.Batch()
 	if err != nil {
@@ -348,7 +352,7 @@ func TestBatching(t *testing.T) {
 	}
 
 	for k, v := range testcases {
-		err := b.Put(datastore.NewKey(k), []byte(v))
+		err := b.Put(ds.NewKey(k), []byte(v))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -360,7 +364,7 @@ func TestBatching(t *testing.T) {
 	}
 
 	for k, v := range testcases {
-		val, err := d.Get(datastore.NewKey(k))
+		val, err := d.Get(ds.NewKey(k))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -376,12 +380,12 @@ func TestBatching(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	err = b.Delete(datastore.NewKey("/a/b"))
+	err = b.Delete(ds.NewKey("/a/b"))
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	err = b.Delete(datastore.NewKey("/a/b/c"))
+	err = b.Delete(ds.NewKey("/a/b/c"))
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -405,6 +409,405 @@ func TestBatching(t *testing.T) {
 		"/f",
 		"/g",
 	}, rs)
+}
+
+func SubtestBasicPutGet(t *testing.T) {
+	d, done := newDS(t)
+	defer done()
+
+	k := ds.NewKey("foo")
+	val := []byte("Hello Datastore!")
+
+	err := d.Put(k, val)
+	if err != nil {
+		t.Fatal("error putting to datastore: ", err)
+	}
+
+	have, err := d.Has(k)
+	if err != nil {
+		t.Fatal("error calling has on key we just put: ", err)
+	}
+
+	if !have {
+		t.Fatal("should have key foo, has returned false")
+	}
+
+	size, err := d.GetSize(k)
+	if err != nil {
+		t.Fatal("error getting size after put: ", err)
+	}
+	if size != len(val) {
+		t.Fatalf("incorrect size: expected %d, got %d", len(val), size)
+	}
+
+	out, err := d.Get(k)
+	if err != nil {
+		t.Fatal("error getting value after put: ", err)
+	}
+
+	if !bytes.Equal(out, val) {
+		t.Fatal("value received on get wasnt what we expected:", out)
+	}
+
+	have, err = d.Has(k)
+	if err != nil {
+		t.Fatal("error calling has after get: ", err)
+	}
+
+	if !have {
+		t.Fatal("should have key foo, has returned false")
+	}
+
+	size, err = d.GetSize(k)
+	if err != nil {
+		t.Fatal("error getting size after get: ", err)
+	}
+	if size != len(val) {
+		t.Fatalf("incorrect size: expected %d, got %d", len(val), size)
+	}
+
+	err = d.Delete(k)
+	if err != nil {
+		t.Fatal("error calling delete: ", err)
+	}
+
+	have, err = d.Has(k)
+	if err != nil {
+		t.Fatal("error calling has after delete: ", err)
+	}
+
+	if have {
+		t.Fatal("should not have key foo, has returned true")
+	}
+
+	size, err = d.GetSize(k)
+	switch err {
+	case ds.ErrNotFound:
+	case nil:
+		t.Fatal("expected error getting size after delete")
+	default:
+		t.Fatal("wrong error getting size after delete: ", err)
+	}
+	if size != -1 {
+		t.Fatal("expected missing size to be -1")
+	}
+}
+
+func TestNotFounds(t *testing.T) {
+	d, done := newDS(t)
+	defer done()
+
+	badk := ds.NewKey("notreal")
+
+	val, err := d.Get(badk)
+	if err != ds.ErrNotFound {
+		t.Fatal("expected ErrNotFound for key that doesnt exist, got: ", err)
+	}
+
+	if val != nil {
+		t.Fatal("get should always return nil for not found values")
+	}
+
+	have, err := d.Has(badk)
+	if err != nil {
+		t.Fatal("error calling has on not found key: ", err)
+	}
+	if have {
+		t.Fatal("has returned true for key we don't have")
+	}
+
+	size, err := d.GetSize(badk)
+	switch err {
+	case ds.ErrNotFound:
+	case nil:
+		t.Fatal("expected error getting size after delete")
+	default:
+		t.Fatal("wrong error getting size after delete: ", err)
+	}
+	if size != -1 {
+		t.Fatal("expected missing size to be -1")
+	}
+}
+
+func SubtestManyKeysAndQuery(t *testing.T) {
+	d, done := newDS(t)
+	defer done()
+
+	var keys []ds.Key
+	var keystrs []string
+	var values [][]byte
+	count := 100
+	for i := 0; i < count; i++ {
+		s := fmt.Sprintf("%dkey%d", i, i)
+		dsk := ds.NewKey(s)
+		keystrs = append(keystrs, dsk.String())
+		keys = append(keys, dsk)
+		buf := make([]byte, 64)
+		_, _ = rand.Read(buf)
+		values = append(values, buf)
+	}
+
+	t.Logf("putting %d values", count)
+	for i, k := range keys {
+		err := d.Put(k, values[i])
+		if err != nil {
+			t.Fatalf("error on put[%d]: %s", i, err)
+		}
+	}
+
+	t.Log("getting values back")
+	for i, k := range keys {
+		val, err := d.Get(k)
+		if err != nil {
+			t.Fatalf("error on get[%d]: %s", i, err)
+		}
+
+		if !bytes.Equal(val, values[i]) {
+			t.Fatal("input value didnt match the one returned from Get")
+		}
+	}
+
+	t.Log("querying values")
+	q := dsq.Query{KeysOnly: true}
+	resp, err := d.Query(q)
+	if err != nil {
+		t.Fatal("calling query: ", err)
+	}
+
+	t.Log("aggregating query results")
+	var outkeys []string
+	for {
+		res, ok := resp.NextSync()
+		if res.Error != nil {
+			t.Fatal("query result error: ", res.Error)
+		}
+		if !ok {
+			break
+		}
+
+		outkeys = append(outkeys, res.Key)
+	}
+
+	t.Log("verifying query output")
+	sort.Strings(keystrs)
+	sort.Strings(outkeys)
+
+	if len(keystrs) != len(outkeys) {
+		t.Fatal("got wrong number of keys back")
+	}
+
+	for i, s := range keystrs {
+		if outkeys[i] != s {
+			t.Fatalf("in key output, got %s but expected %s", outkeys[i], s)
+		}
+	}
+
+	t.Log("deleting all keys")
+	for _, k := range keys {
+		if err := d.Delete(k); err != nil {
+			t.Fatal(err)
+		}
+	}
+}
+
+// Tests from basic_tests from go-datastore
+func TestBasicPutGet(t *testing.T) {
+	d, done := newDS(t)
+	defer done()
+
+	k := ds.NewKey("foo")
+	val := []byte("Hello Datastore!")
+
+	err := d.Put(k, val)
+	if err != nil {
+		t.Fatal("error putting to datastore: ", err)
+	}
+
+	have, err := d.Has(k)
+	if err != nil {
+		t.Fatal("error calling has on key we just put: ", err)
+	}
+
+	if !have {
+		t.Fatal("should have key foo, has returned false")
+	}
+
+	out, err := d.Get(k)
+	if err != nil {
+		t.Fatal("error getting value after put: ", err)
+	}
+
+	if !bytes.Equal(out, val) {
+		t.Fatal("value received on get wasnt what we expected:", out)
+	}
+
+	have, err = d.Has(k)
+	if err != nil {
+		t.Fatal("error calling has after get: ", err)
+	}
+
+	if !have {
+		t.Fatal("should have key foo, has returned false")
+	}
+
+	err = d.Delete(k)
+	if err != nil {
+		t.Fatal("error calling delete: ", err)
+	}
+
+	have, err = d.Has(k)
+	if err != nil {
+		t.Fatal("error calling has after delete: ", err)
+	}
+
+	if have {
+		t.Fatal("should not have key foo, has returned true")
+	}
+	SubtestBasicPutGet(t)
+}
+
+func TestManyKeysAndQuery(t *testing.T) {
+	d, done := newDS(t)
+	defer done()
+
+	var keys []ds.Key
+	var keystrs []string
+	var values [][]byte
+	count := 100
+	for i := 0; i < count; i++ {
+		s := fmt.Sprintf("%dkey%d", i, i)
+		dsk := ds.NewKey(s)
+		keystrs = append(keystrs, dsk.String())
+		keys = append(keys, dsk)
+		buf := make([]byte, 64)
+		_, _ = rand.Read(buf)
+		values = append(values, buf)
+	}
+
+	t.Logf("putting %d values", count)
+	for i, k := range keys {
+		err := d.Put(k, values[i])
+		if err != nil {
+			t.Fatalf("error on put[%d]: %s", i, err)
+		}
+	}
+
+	t.Log("getting values back")
+	for i, k := range keys {
+		val, err := d.Get(k)
+		if err != nil {
+			t.Fatalf("error on get[%d]: %s", i, err)
+		}
+
+		if !bytes.Equal(val, values[i]) {
+			t.Fatal("input value didnt match the one returned from Get")
+		}
+	}
+
+	t.Log("querying values")
+	q := dsq.Query{KeysOnly: true}
+	resp, err := d.Query(q)
+	if err != nil {
+		t.Fatal("calling query: ", err)
+	}
+
+	t.Log("aggregating query results")
+	var outkeys []string
+	for {
+		res, ok := resp.NextSync()
+		if res.Error != nil {
+			t.Fatal("query result error: ", res.Error)
+		}
+		if !ok {
+			break
+		}
+
+		outkeys = append(outkeys, res.Key)
+	}
+
+	t.Log("verifying query output")
+	sort.Strings(keystrs)
+	sort.Strings(outkeys)
+
+	if len(keystrs) != len(outkeys) {
+		t.Fatalf("got wrong number of keys back, %d != %d", len(keystrs), len(outkeys))
+	}
+
+	for i, s := range keystrs {
+		if outkeys[i] != s {
+			t.Fatalf("in key output, got %s but expected %s", outkeys[i], s)
+		}
+	}
+
+	t.Log("deleting all keys")
+	for _, k := range keys {
+		if err := d.Delete(k); err != nil {
+			t.Fatal(err)
+		}
+	}
+
+	SubtestManyKeysAndQuery(t)
+}
+
+func TestTxn(t *testing.T) {
+	d, done := newDS(t)
+	defer done()
+	txn, err := d.NewTransaction(false)
+	if err != nil {
+		t.Fatal(err)
+	}
+	testKey := ds.NewKey("helloworld")
+	failKey := ds.NewKey("donothave")
+	testValue := []byte("hello world")
+	if err := txn.Put(testKey, testValue); err != nil {
+		t.Fatal(err)
+	}
+	if err := txn.Delete(testKey); err != nil {
+		t.Fatal(err)
+	}
+	if err := txn.Put(testKey, testValue); err != nil {
+		t.Fatal(err)
+	}
+	if _, err := txn.Query(dsq.Query{}); err == nil {
+		t.Fatal("error expected")
+	}
+	if size, err := txn.GetSize(testKey); err != nil {
+		t.Fatal(err)
+	} else if size != len(testValue) {
+		t.Fatalf("bad size, got %v, wanted %v", size, len(testValue))
+	}
+	if has, err := txn.Has(testKey); err != nil {
+		t.Fatal(err)
+	} else if !has {
+		t.Fatal("should have key")
+	}
+	if has, err := txn.Has(failKey); err != nil {
+		t.Fatal(err)
+	} else if has {
+		t.Fatal("should not have key")
+	}
+	if val, err := txn.Get(testKey); err != nil {
+		t.Fatal(err)
+	} else if string(val) != string(testValue) {
+		t.Fatal("bad value returned")
+	}
+	if _, err := txn.Get(failKey); err != ds.ErrNotFound {
+		t.Fatal("bad error returned")
+	}
+	if err := txn.Commit(); err != nil {
+		t.Fatal(err)
+	}
+	if err := txn.Commit(); err == nil {
+		t.Fatal("error expected")
+	}
+	txn.Discard()
+}
+
+func TestSuite(t *testing.T) {
+	d, done := newDS(t)
+	defer done()
+
+	dstest.SubtestAll(t, d)
 }
 
 func expectKeyFilterMatches(t *testing.T, actual dsq.Results, expect []string) {
